+++ hyperspy/external/mpfit/mpfitexpr.py
