+++ hyperspy/external/mpfit/tests/test_mpfit.py
@@ -2,7 +2,7 @@ from numpy.testing import *
 import numpy as N
 import copy
 
-from hyperspy.misc.mpfit.mpfit import mpfit
+from hyperspy.external.mpfit.mpfit import mpfit
 
 
 def Flin(x, p):
