+++ setup.py
@@ -161,8 +161,8 @@ with update_version_when_dev() as version:
                   'hyperspy.misc.eds',
                   'hyperspy.misc.io',
                   'hyperspy.misc.machine_learning',
-                  'hyperspy.misc.mpfit',
-                  'hyperspy.misc.mpfit.tests',
+                  'hyperspy.external.mpfit',
+                  'hyperspy.external.mpfit.tests',
                   'hyperspy.external',
                   'hyperspy.external.astroML',
                   ],
